day3 completed
Wow I'm getting more and more jaded with these questions. At first I was
all like "let's think of the smartest time and space efficient
algorithms to do this!" only to slowly realize that no, it's not really
possible.  These questions seem to be DESIGNED to only attract the brute
force solutions. But maybe that's not so bad? It helps you get familiar
with the brute force methods available in your programming language.

For example, julia has the 'findall' function which made 3b laughably
trivial. Look at this line!

    # counts all the cells on the checkerboard belonging to 'id'
    actualcells = length(findall(a->a==id, checkerboard))

I could have inlined this shit even more if I wanted to.
Expression-based programming syntax is great