Got some more headway on 4.jl
I think I may need to implement a struct that will hold the current
status at each iteration of the strarraysorted loop. Also I gave in to
the dark side and started writing C-style one-use variables as flags for
further control flow. It's getting pretty ugly, but I guess I'll have to
get used to it.

I also realized that I like functional style programming not because
it's superior to imperative but because I'm afraid of pure imperative
code. It's hard to read and understand. I'm also beginning to understand
why programmers of old developed Object Oriented Programming(even though
it gets a lot of flak now).