day5a completed, but bout to refactor it
Right now doing a naive pass over the string each time, passing over the
string repeatedly until the output length is same as the input length.
This method is S L O W. It takes about a minute for it to complete.

I think it would be a lot faster if I extracted the last character from
the IOBuffer if the incoming character matches the last character
inserted. But this hinges on the fact whether it is possible to 'pop'
characters out of the IOBuffer. More reading up on documentation I
suppose